requirements.txt
aiofiles
aiohttp
cachetools
discord.py
matplotlib
requests
ruamel.yaml
